:white_check_mark: test: add test for the new api, coverage 100%
